% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eml.R
\name{check_eml}
\alias{check_eml}
\title{Run checks on EML}
\usage{
check_eml(directory = here::here())
}
\arguments{
\item{directory}{path to the metadata file. Defaults to the current working directory. Make sure there is only one .xml file in the directory.}
}
\description{
Runs a series of checks on EML metadata based on functions in \verb{r DPchecker::run_congruence_checks()}.
}
\examples{
\dontrun{
check_eml()
}
}
