% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{combine_geographic_coverage}
\alias{combine_geographic_coverage}
\title{Creates a geographic coverage text file for a data package with location data
in multiple tables.}
\usage{
combine_geographic_coverage(
  path = getwd(),
  data_path = path,
  tables,
  latitude,
  longitude,
  sitename,
  write = TRUE
)
}
\arguments{
\item{path}{defaults to the working directory. This is the path to  which the
geography text file will be written.}

\item{data_path}{defaults to the argument passed to path. This is the path to
the tables with location data.}

\item{tables}{is a character vector containing the table names (including the
".csv" ending) with location data.}

\item{latitude}{is a character vector containing the column names with decimal
latitude data in each table. It must be the same length as the tables argument
and in the same order.}

\item{longitude}{is a character vector containing the column names with decimal
longitude data in each table. It must be the same length as the tables argument
and in the same order.}

\item{sitename}{is a character vector containing the column names with site name
or site identifier data in each table. It must be the same length as the tables
argument and in the same order.}

\item{write}{defaults to TRUE. A logical value that indicates whether or not the
geographic coverage table will be written as a tab delimited text file to the
specified path.}
}
\description{
\code{combine_geographic_coverage} is a wrapper around EMLassemblyline::
template_geographic_coverage, but allows multiple files with location data as
inputs (much like EMLassemblyline::template_taxonomic_coverage). It writes a
csv file with all location data to a temporary directory, then uses that csv
as an input for EMLassemblyline::template_geographic_coverage to create the
geographic coverage text file. The temporary directory is cleared when the R
session ends, so the user doesn't have to remove the csv file and it does not
interfere with the make_eml and run_congruence_checks functions.
}
\examples{
\dontrun{
combine_geographic_coverage(path = "data/2313987",
                            tables = c("Sites.csv", "Subsites.csv"),
                            latitude = c("Latitude", "Latitude"),
                            longitude = c("Longitude", "Longitude"),
                            sitename = c("SiteCode", "SubsiteCode"),
                            write = TRUE)
}
}
