% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_user_input3}
\alias{.get_user_input3}
\title{Get open ended user input}
\usage{
.get_user_input3()
}
\value{
character, typically 1, 2, or 3 but could be any character string.
}
\description{
Does not prompt for user input. Takes any user input supplied and returns it.
}
\examples{
\dontrun{
var1 <- .get_user_input3()
}
}
