% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.set_npspublisher}
\alias{.set_npspublisher}
\title{inject NPS Publisher info into metadata}
\usage{
.set_npspublisher(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
eml_object
}
\description{
.set_npspublisher injects static NPS-specific publisher info into eml documents. Calls the sub-function set.forOrByNPS, which adds an additionalMetadata element with for or by NPS = TRUE.
}
\details{
checks to see if the publisher element exists, and if not injects NPS-specific info into EML such as publisher, publication location, and ROR id - the types of things that will be the same for all NPS data or non-data publications and do not require user input. This function will be embedded in all set. and write. class functions (and get. functions?).
}
\examples{
 \dontrun{
.set_npspublisher(eml_object)
}
}
