% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eml_get_simpleNPS}
\alias{eml_get_simpleNPS}
\title{alias for eml_get_simple}
\usage{
eml_get_simpleNPS(...)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{elment}{is the name of the element to be extracted. If multiple occurances are found, will extract all.}
}
\description{
alias for eml_get_simple
}
\details{
eml_get_simpleNPS serves as an alias for the arcticdatautils function eml_get_simple. This means the eml_get_simple() function will be called using current local version of arcticdatautils on the users's machine rather than whatever version was available when the package was built.
}
