% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eml_getNPS}
\alias{eml_getNPS}
\title{alias for eml_get}
\usage{
eml_getNPS(...)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{elment}{is the name of the element to be extracted. If multiple occurrences are found, will extract all.}
}
\description{
alias for eml_get
}
\details{
eml_getNPS serves as an alias for the EML function eml_get. This means the eml_get() function will be called using current local version of EML on the users's machine rather than whatever version was available when the package was built.
}
