% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMLeditor-package.R
\docType{package}
\name{EMLeditor-package}
\alias{EMLeditor}
\alias{EMLeditor-package}
\title{EMLeditor: View and Edit EML Metadata}
\description{
This package will be of most use to the U.S. National Park Service data scientists and managers seeking to generate EML-formatted metadata for datapackages. EML-formatted .xml files are typically constructed using EDI's EMLassemblyline package and then imported as an R-object using the EML package. EMLeditor allows the user to view the contents of the R object and add/edit aspects of metadata crucial for publication in the U.S. National Park Service DataStore repository. For instance, a user can view and edit a DOI, a link to a DRR, Park Unit connections, information about Confidential Unclassified Information (CUI), and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/doi-nps/EMLeditor}
  \item \url{https://doi-nps.github.io/EMLeditor/}
  \item Report bugs at \url{https://github.com/doi-NPS/EMLeditor/issues}
}

}
\author{
\strong{Maintainer}: Robert Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})

Authors:
\itemize{
  \item Robert Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})
  \item Judd Patterson \email{Judd_Patterson@nps.gov} (\href{https://orcid.org/0000-0002-0951-7917}{ORCID})
}

Other contributors:
\itemize{
  \item Issac Quevedo (\href{https://orcid.org/0000-0003-0129-981X}{ORCID}) [contributor]
  \item Melina Kompella (\href{https://orcid.org/0009-0000-3730-8082}{ORCID}) [contributor]
}

}
\keyword{internal}
