% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get.beginDate}
\alias{get.beginDate}
\title{returns the first date}
\usage{
get.beginDate(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a text string
}
\description{
get.beginDate returns the date of the earliest data point in the datapackage
}
\details{
returns the date from the <beginDate> tag. Although dates should be formatted according to ISO-8601 (YYYY-MM-DD) it will also check for a few other common formats and return the date as a text string: "DD Month YYYY"
}
\examples{
get.beginDate(emlObject)
}
