% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get.DOI}
\alias{get.DOI}
\title{returns the DOI}
\usage{
get.DOI(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a text string
}
\description{
returns a text string that is the DOI for the datapackage
}
\details{
accesses the contents of the<alternateIdentifier> tag and does some text manipulation to return a string with the DOI including the URL and prefaced by 'doi: '.
}
