% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get.lit}
\alias{get.lit}
\title{Get literature cited}
\usage{
get.lit(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
character string
}
\description{
get.lit prints bibtex fromated literature cited to the screen.
}
\details{
get.lit currently only supports bibtex formatted references. get.lit gets items from the <literatureCited> tag and prints them to the screen.
}
\examples{
\dontrun{
get.lit(emlObject)
}
}
