% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get.producingUnits}
\alias{get.producingUnits}
\title{Returns the Producing Units}
\usage{
get.producingUnits(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a character sting
}
\description{
get.producingUnits returns whatever is in the metadataProvider eml element. Set this to the producing units using the set.producingUnits function.
}
\examples{
\dontrun{
get.producingUnits(emlObject)
}
}
