% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.unitPolygon}
\alias{get.unitPolygon}
\title{Get Park Unit Polygon}
\usage{
get.unitPolygon(Unit_Code)
}
\arguments{
\item{Unit_Code}{a string (typically 4 characters) that is the park unit code.}
}
\value{
a park polygon
}
\description{
get.unitPolygon gets the polygon for a given park unit.
}
\details{
retrieves a geoJSON string for a polygon of a park unit from NPS Rest services. Note: This is not the official boundary (erm... ok then what is it?!?).
}
\examples{
 \dontrun{
poly<-get.unitPolygon("BICY")
}
}
