% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_abstract}
\alias{get_abstract}
\title{returns the abstract}
\usage{
get_abstract(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
a text string
}
\description{
returns the text from the \if{html}{\out{<abstract>}} tag.
}
\details{
get_abstract returns the text from the \if{html}{\out{<abstract>}} tag and attempts to clean up common text issues, such as enforcing UTF-8 formatting, getting rid of carriage returns, new lines, \if{html}{\out{<para>}} and \if{html}{\out{<literalLayout>}} tags and mucks about with layout, line breaks, etc. IF you see characters you don't like in the abstract, make sure to edit your abstract in a text editor (e.g. Notepad and NOT a Word). You should save the text to a new object and view it using writeLines()
}
\examples{
\dontrun{
abstract <- get_abstract(eml_object)
writeLines(abstract)
}
}
