% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_additional_info}
\alias{get_additional_info}
\title{Get additional information (Notes on DataStore)}
\usage{
get_additional_info(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
List
}
\description{
\code{get_methods()} returns the text stored in the methods element of EML metadata. The returned text is not manipulated in any way. DataStore unlist the returned object (get rid of tags such as $methodStep, $methodStep$description and $methodStep$description$para and remove the numbers in brackets). the "\\n" character combination is interpretted as a line break (as are blank lines). However, DataStpre will not filter out stray characters such as &#13;. Use the \code{set_methods()} function to edit and replace the text stored in the methods field.
}
\examples{
\dontrun{
get_additional_info(eml_object)
}
}
