% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_catvar_tables}
\alias{get_catvar_tables}
\title{Creates categorical variable tables from an EML object}
\usage{
get_catvar_tables(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file)
using EML::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
a nested table with one catvar table for each data table in the EML file
}
\description{
\code{get_catvar_tables} takes an EML object and returns a nested table of
all the categorical variable tables pulled from the metadata, using EML::get_attributes()
}
\examples{
\dontrun{
get_catvar_tables(example_EML)
}
}
