% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_cui_code}
\alias{get_cui_code}
\title{returns a CUI dissemination code statement}
\usage{
get_cui_code(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
a text string
}
\description{
\code{get_cui_code()} returns an English-language translation of the CUI dissemination codes. It supersedes \code{get_cui()}, which has been deprecated.
}
\details{
\code{get_cui_code()} accesses the contents of the Controlled Unclassified Information (CUI) tag, \if{html}{\out{<CUI>}} and returns an appropriate string of english-language text based on the properties of the CUI code. If thee \if{html}{\out{<CUI>}} tag is empty or does not exist, get_cui alerts the user and suggests specifying CUI using the set_cui() funciton.
}
\examples{
\dontrun{
get_cui(eml_object)
}
}
