% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_cui_marking}
\alias{get_cui_marking}
\title{Returns the CUI marking}
\usage{
get_cui_marking(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
String (invisibly)
}
\description{
For data with controlled unclassified information (CUI), \code{get_cui_marking()} eturns the specific marking and the english language explanation of the marking. For data without CUI, it informs that there is no CUI and returns the code "PUBLIC".
}
\details{
CUI markings are defined by the U.S. National Archives (nara.gov). NPS users can designate one of three CUI markings, plus the code "PUBLIC" (essentially, no marking necessary). The three markings are: SP-NPSR, SP-HISTP or SP-ARCHR.
For more information on CUI markings, please visit the \href{https://www.archives.gov/cui/registry/category-marking-list}{CUI Markings} list maintained by the National Archives.
}
\examples{
\dontrun{
get_cui_marking(eml_object)
}
}
