% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_doi}
\alias{get_doi}
\title{returns the DOI}
\usage{
get_doi(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
a text string
}
\description{
returns a text string that is the DOI for the data package
}
\details{
\code{get_doi()} accesses the contents of the\if{html}{\out{<alternateIdentifier>}} tag and does some text manipulation to return a string with the DOI including the URL and prefaced by 'doi: '.
}
\examples{
\dontrun{
get_doi(eml_object)
}
}
