% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_drr_doi}
\alias{get_drr_doi}
\title{returns the DOI of the associated DRR}
\usage{
get_drr_doi(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a text string
}
\description{
get_drr_doi returns a text string with the associated Data Release Report (DRR)'s DOI.
}
\details{
get_drr_doi accesses the <usageCitation> tag(s) and searches an alternateIdentifier tag. If that element is found, the contents of that element are returned. If the title element is empty or not present, the user is warned and pointed to the set_drr function to add the DOI of an associated DRR.
}
\examples{
\dontrun{
get_drr_doi(eml_object)
}
}
