% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_drr_title}
\alias{get_drr_title}
\title{returns the title of the associated DRR}
\usage{
get_drr_title(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a text string
}
\description{
get_drr_title returns a text string with the associated Data Release Report (DRR)'s Title.
}
\details{
get_drr_title accesses useageCitation under dataset and returns the title element, if it is found. If it is not found, the user is warned and pointed ot the set_drr function to add the title of the associated DRR.
}
\examples{
\dontrun{
 get_drr_title(eml_object)
 }
}
