% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_eml_simple}
\alias{get_eml_simple}
\title{Returns a simple list of eml elements}
\usage{
get_eml_simple(eml_object, eml_element)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}

\item{eml_element}{is a string indicating the element within the eml_object that should be accessed and returned.}
}
\value{
a list of values
}
\description{
This function wraps EML::get_eml and returns the requested object in a much simpler format. It works best with eml elements that do not have child elements.
}
\examples{
\dontrun{
get_eml_simple(eml_object, "geographicDescription")
}
}
