% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_end_date}
\alias{get_end_date}
\title{returns the last date}
\usage{
get_end_date(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a text sting
}
\description{
get_end_date returns the date of the last data point in the data package
}
\details{
returns the date from the <endDate> tag. Although dates should be formatted according to ISO-8601 (YYYY-MM-DD) it will also check a few other common formats and return the date as a text string: "DD Month YYYY"
}
\examples{
 \dontrun{
get_end_date(emlObject)
}
}
