% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_title}
\alias{get_title}
\title{returns the data package title}
\usage{
get_title(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
a text string
}
\description{
get_title returns a text string that is the title of the data package
}
\examples{
\dontrun{
get_title(eml_object)
}
}
