% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_interactions.R
\name{remove_datastore_files}
\alias{remove_datastore_files}
\title{Remove files from a data package Reference on DataStore}
\usage{
remove_datastore_files(data_store_reference, force = FALSE, dev = FALSE)
}
\arguments{
\item{data_store_reference}{Integer. the 7-digit DataStore Reference number}

\item{force}{Logical. Defaults to FALSE. Set to TRUE to avoid interactive components and most user feedback.}

\item{dev}{Logical. Defaults to FALSE. set to TRUE if you want to detete files from a DataStore reference on the Development server.}
}
\description{
The \code{remove_datastore_files()} function requires that you are logged in to the VPN. You must also have permissions to edit the DataStore Reference in question, and the Reference must not be activated. In that case, \code{remove_datastore_files()} detaches all files associated with the relevant DataStore Reference. Once the files are detached, they cannot be re-attached.  Instead, updated files can be re-uploaded and attached to the reference via \code{upload_data_package()}.

In the interactive mode (force = FALSE), you will be informed of the Reference number, Reference title, Reference creation date, and a list of files currently attached to the Reference. Once the files are successfully detached, you will be informed of a list of file names that were detached.

Note that \code{remove_datastore_files()} is intended to be used with the data package Reference type on DataStore but in theory will allow you to remove files from ANY reference type.
}
\examples{
\dontrun{
remove_datastore_files(1234567)
remove_datastore_files(1234567, force = TRUE, dev = TRUE)
}
}
