% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_interactions.R
\name{reserve_doi}
\alias{reserve_doi}
\title{Reserve DOI}
\usage{
reserve_doi(reference_type = c("data_package", "tabular data", "DRR"), title)
}
\arguments{
\item{reference_type}{a string indicating the type of reference to use. The
default is data_package}

\item{title}{a string corresponding to the title for the draft reference
(this can be edited later if need be)}
}
\value{
String. a 7-digit string corresponding to the draft Reference ID.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Initiates a draft reference on DataStore (https://irma.nps.gov/DataStore/).
The draft reference number is stored and used to generate a DOI, which can
be inserted into metadata. Once the draft reference is created, the DOI has
been reserved and will not be available to others. However, it has not yet
been submitted to the DOI registry system (which will happen after the data
package has been reviewed), nor is it active in the DOI registry system. The
reservation is internal to NPS.
}
\details{
You must be logged on to the VPN or in an NPS office to initiate a
draft reference on DataStore. It is important that your data and metadata are
uploaded to the correct draft reference. While the reference is in draft
status, you can make changes to the reference including file uploads. Do NOT
activate the reference at this stage! Once a reference is activated you will
not be able to make further changes without creating a new version.
}
\examples{
\dontrun{
reserve_doi(reference_type)
}
}
