% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set.abstract}
\alias{set.abstract}
\title{adds an abstract}
\usage{
set.abstract(emlObject, abstract, NPS = TRUE)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EML::read_eml(<filename>, from="xml").}

\item{abstract}{is a text string that is your abstract. You can generate this directly in R or import a .txt file.}

\item{NPS}{defaults to TRUE. Checks EML for NPS publisher info and injects it if publisher is empty. If publisher already exists, does nothing. If you are not publishing with NPS, set to FALSE}
}
\value{
an EML-formatted R object
}
\description{
set.abstract adds (or replaces) a simple abstract.
}
\details{
checks for an abstract. If no abstract is found, it inserts the abstract given in @param abstract. If an existing abstract is found, the user is asked whether they want to replace it or not and the appropriate action is taken. Currently set.abstract does not allow for paragraphs or complex formatting. You are strongly encouraged to open your abstract in a text editor such as notepad and make sure there are no stray characers. If you need multiple paragraphs, you will need to do that via EMLassemblyline (for now).
}
\examples{
set.abstract(emlObject "This is a very short abstract")
}
