% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set.CUI}
\alias{set.CUI}
\title{Adds CUI to metadata}
\usage{
set.CUI(emlObject, CUI)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{CUI}{a string consisting of one of 5 potential CUI codes. Pay attention to the spaces:
FED ONLY - Contains CUI. Only federal employees should have access (similar to "internal only" in DataStore)
FEDCON - Contains CUI. Only federal employees and federal contractors should have access (also very much like current "internal only" setting in DataStore)
DL ONLY - Contains CUI. Should only be available to a names list of individuals (where and how to list those individuals TBD)
NOCON - Contains  CUI. Federal, state, local, or tribal employees may have access, but contractors cannot.
PUBVER - Does NOT contain CUI. The original data contained CUI, but in this datapackage CUI have been obscured so that it no longer contains CUI.
PUBFUL - Does NOT contain CUI. The original data contained no CUI. No data were obscured or altered to generate the data package}
}
\value{
an EML-formatted R object
}
\description{
set.CUI adds CUI codes to EML metadata
}
\details{
set.CUI adds a CUI code to the tag <CUI> under <additionalMetadata><metadata>.
}
\examples{
set.CUI(emlObject, "PUBFUL")
}
