% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set.DRRdoi}
\alias{set.DRRdoi}
\title{adds DRR connection}
\usage{
set.DRRdoi(emlObject, DRRrefID, NPS = TRUE)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{DRRrefID}{a 7-digit string that is the DataStore Reference ID for the DRR associated with the data package.}

\item{NPS}{defaults to TRUE. Checks EML for NPS publisher info and injects it if publisher is empty. If publisher already exists, does nothing. If you are not publishing with NPS, set to FALSE}
}
\value{
an EML-formatted R object
}
\description{
set.DRRdoi adds the DOI of an associated DRR
}
\details{
adds uses the DataStore Reference ID for an associate DRR to the <useageCitation> as a properly formatted DOI (prefaced with "DRR: ") to the <useageCitation> tag. ####CAUTION: in the current implimentation this may overwrite any other useageCitation info.
}
\examples{
set.DRRdoi(emlObject, "2293234")
}
