% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set.producingUnits}
\alias{set.producingUnits}
\title{Sets Producing Units for use in DataStore}
\usage{
set.producingUnits(emlObject, prodUnit, NPS = TRUE)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{prodUnit}{A string that is the producing unit Unit Code}

\item{NPS}{defaults to TRUE. Checks EML for NPS publisher info and injects it if publisher is empty. If publisher already exists, does nothing. If you are not publishing with NPS, set to FALSE}
}
\value{
an EML object
}
\description{
inserts the unit code for the producing unit for the data/metadata into the EML metdata file.
}
\details{
inserts the unit code into the metadataProvider element. Currently cannot add to existing metadataProvider fields; it will just over-write them. It also currently only handles a single producing unit.
}
\examples{
\dontrun{
set.producingUnits(emlObject, "ROMO")
}
}
