% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set.protocol}
\alias{set.protocol}
\title{Adds a connection to the protocol under which the data were collected}
\usage{
set.protocol(emlObject, protocolID, NPS = TRUE)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{protocolID}{a string. The 7-digit number identifying the protocol in DataStore under which the data were collected.}

\item{NPS}{Logical. Checks EML for NPS publisher info and injects it if publisher is empty. If publisher already exists, does nothing. If you are not publishing with NPS, set to FALSE. If NPS=TRUE, the originatingAgency will be set to NPS and the field that maps to DataStore's "by or for NPS" will be set to TRUE.}
}
\description{
set.protocol adds a metadata link to the protocol under which the data being described were collected. It automatically inserts a link to the DataStore landing page for the protocol as well as ?????
}
\details{
set.protocol requires that you have your protocols and projects organized in a specific fashion in DataStore. Errors generated by this function my stem from either a protocol that has not been published (or is not publicly available) or an obsolete protocol/project organization within DataStore.
}
\examples{
\dontrun{
set.protocol(emlObject, 2222140)
}

}
