% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.version}
\alias{set.version}
\title{Add EMLeditory version}
\usage{
set.version(emlObject)
}
\arguments{
\item{emlObject}{}
}
\description{
set.version adds the current version of EMLeditor to the metadata
}
\details{
set.version adds the current version of EMLeditor to the metadata, specifically in the
}
\examples{
set.version(emlObject)
}
