% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set_park_units}
\alias{set_park_units}
\title{Add Park Unit Connections to metadata}
\usage{
set_park_units(eml_object, park_units, NPS = TRUE)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{park_units}{a list of comma-separated strings where each string is a park unit code.}

\item{NPS}{defaults to TRUE. Checks EML for NPS publisher info and injects it if publisher is empty. If publisher already exists, does nothing. If you are not publishing with NPS, set to FALSE}
}
\value{
an EML-formatted R object
}
\description{
set_park_units adds all specified park unit connections and their N, E, S, W bounding boxes to <geographicCoverage>.
}
\details{
Adds the Park Unit Connection(s) to a <coverage>. Park Unit Connection(s) are the (typically) four-letter codes describing the park unit(s) where data were collected (e.g. ROMO, not ROMN). Each park unit connection is given a separate <geographicCoverage> element. For each park unit connection, the unit name will be listed under <geographicDescription> and prefaced with "NPS Unit Connections:". Required child elements (bounding coordinates) are auto populated. If other <geographicCoverage> elements exist, set_park_units will add to them, not overwrite them. If not other <geographicCoverage> elements exist, set_park_units will create a new set of <geographicCoverage> elements.
}
\examples{
 \dontrun{
park_units<-("ROMO, GRSD, TRYME")
set_park_units(eml_object, park_units)
}
}
