% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{write_catvar_tables}
\alias{write_catvar_tables}
\title{Writes categorical variable tables from an EML object as text files}
\usage{
write_catvar_tables(eml_object, path = here::here())
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file)
using EML::read_eml(\if{html}{\out{<filename>}}, from="xml").}

\item{path}{default to the working directory. This path determines where the
txt files will be written.}
}
\description{
\code{write_catvar_tables} is a wrapper around get_catvar_tables().
It takes an EML object and writes one categorical variable table for each data table
in the EML file. It writes the attribute tables as .txt files in the working directory
or a specified path. This function is useful for recreating metadata with EMLassemblyline,
when the user doesn't have access to the original text files.
}
\examples{
\dontrun{
write_catvar_tables(example_EML)
}
}
