% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eml.R
\name{write_readme}
\alias{write_readme}
\title{Writes a README file}
\usage{
write_readme(eml_object, outfile = "")
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EML::read_eml(<filename>, from="xml").}

\item{outfile}{is the name of the file you want to write, typically *.txt.}
}
\value{
a character sting in readable format (saved to the given outfile)
}
\description{
write_readme writes a readme file based on the current metadata
}
\details{
write_readme writes a mock-up of the readme file that will eventually be automatically generated by DataStore. This file is for error checking purposes only. If something looks off, you can go back and fix your metadata to correct it but you should not upload this readme file with your data package.
}
\examples{
\dontrun{
write_readme(eml_object, "TestReadMe.txt")
}
}
