% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eml.R
\name{check_eml}
\alias{check_eml}
\title{Run checks on EML}
\usage{
check_eml(path = here::here())
}
\arguments{
\item{path}{to the metadata file. Defaults to the current working directory. Make sure there is only one .xml file in the directory.}
}
\value{
message
}
\description{
Runs a series of checks on EML metadata based on functions in DPchecker's run_congruence_checks()`.
}
\examples{
\dontrun{
check_eml()
}
}
