% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_unit_polygon}
\alias{.get_unit_polygon}
\title{Get Park Unit Polygon}
\usage{
.get_unit_polygon(unit_code)
}
\arguments{
\item{unit_code}{a string (typically 4 characters) that is the park unit code.}
}
\value{
a park polygon
}
\description{
.get_unit_polygon gets the polygon for a given park unit.
}
\details{
retrieves a geoJSON string for a polygon of a park unit from NPS Rest services. Note: This is not the official boundary (erm... ok then what is it?!?).
}
\examples{
\dontrun{
poly <- .get_unit_polygon("BICY")
}
}
