% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.set_version}
\alias{.set_version}
\title{Add/update EMLeditor version}
\usage{
.set_version(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
eml_object
}
\description{
.set_version adds the current version of EMLeditor to the EML document.
}
\details{
.set_version adds the current version of EMLeditor to the metadata, specifically in the "additionalMetadata" element
}
\examples{
\dontrun{
.set_version(eml_object)
}
}
