% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_additional_info}
\alias{get_additional_info}
\title{Get additional information (Notes on DataStore)}
\usage{
get_additional_info(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
String
}
\description{
\code{get_additional_info()} returns the text in the additionalInformation element of EML. This text will be used to populate the "Notes" sectionon the DataStore reference page. There is no strict limit on what can and cannot go in to the additionalInformation/Notes section. However, DataStore will not filter out stray characters such as &#13;. Use the \code{set_additional_info()} function to edit and replace the text stored in the additionalInformation (notes) field.
}
\examples{
\dontrun{
get_additional_info(eml_object)
}
}
