% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_cui}
\alias{get_cui}
\title{returns a CUI statement}
\usage{
get_cui(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a text string
}
\description{
get_cui returns an english-language translation of the CUI codes
}
\details{
get_cui accesses the contents of the Controlled Unclassified Information (CUI) tag, <CUI> and returns an appropriate string of english-language text based on the properties of the CUI code. If thee <CUI> tag is empty or does not exist, get_cui alerts the user and suggests specifying CUI using the set_cui() funciton.
}
\examples{
 \dontrun{
get_cui(emlObject)
}
}
