% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_drr_doi}
\alias{get_drr_doi}
\title{returns the DOI of the associated DRR}
\usage{
get_drr_doi(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}
}
\value{
a text string
}
\description{
get_drr_doi returns a text string with the associated Data Release Report (DRR)'s DOI.
}
\details{
get_drr_doi accesses the <usageCitation> tag(s) and searches for the string "DRR: https://doi.org/". If that string is found, the contents of that tag are returned. If the <usageCitation> tag is empty or not present, the user is informed and pointed to the set_drr_doi() function to add the DOI of an associated DRR.
}
\examples{
 \dontrun{
get_drr_doi(emlObject)
}
}
