% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMLfunctions.R
\name{get_producing_units}
\alias{get_producing_units}
\title{Returns the Producing Units}
\usage{
get_producing_units(eml_object)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(\if{html}{\out{<filename>}}, from="xml").}
}
\value{
a character sting
}
\description{
get_producing_units returns whatever is in the metadataProvider eml element. Set this to the producing units using the set_producing_units function.
}
\examples{
\dontrun{
get_producing_units(eml_object)
}
}
