% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set.DOI}
\alias{set.DOI}
\title{Check & set a DOI}
\usage{
set.DOI(emlObject, DSref, NPS = TRUE)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{DSref}{is the same as the 7-digit reference code generated on DataStore when a draft reference is initiated.}

\item{NPS}{is a logical that defaults to TRUE. Checks EML for NPS publisher info and injects it if publisher is empty. If publisher already exists, does nothing. If you are not publishing with NPS, set to FALSE}
}
\value{
an EML-formatted R object
}
\description{
Check & set a DOI
}
\details{
This function  checks to see if there is a DOI in the <alternateIdentifier> tag. The EMLassemblyline package stores data package DOIs in this tag (although the official EML schema has the DOI in a different location). If there is no DOI in the <alternateIdentifier> tag, the function adds a DOI & reports the new DOI. If there is a DOI, the function reports the existing DOI, and prompts the user for input to either retain the existing DOI or overwrite it. Reports back the existing or new DOI, depending on the user input..
}
\examples{
 \dontrun{
 emlObject<-set.DOI(emlObject, DSref)
 }
}
