% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.forByNPS}
\alias{set.forByNPS}
\title{Set "For or By" NPS}
\usage{
set.forByNPS(emlObject)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EML::read_eml(<filename>, from="xml").}
}
\value{
emlObject
}
\description{
set.forByNPS adds an element to additionalMetadata with For or By NPS set to TRUE and a second element agencyOriginated set to "NPS" with the understanding that all data products created for or by the NPS have NPS as the originating agency.
}
\examples{
 \dontrun{
set.forByNPS(emlObject)
}
}
