% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set.parkUnits}
\alias{set.parkUnits}
\title{Adds Park Unit Connections to metadata}
\usage{
set.parkUnits(emlObject, ParkUnits)
}
\arguments{
\item{emlObject}{is an R object imported (typically from an EML-formatted .xml file) using EmL::read_eml(<filename>, from="xml").}

\item{ParkUnits}{a string of comma-separated park unit codes}
}
\value{
an EML-formatted R object
}
\description{
Adds Park Unit Connections to metadata
}
\details{
Add the Park Unit Connection(s) to a <geographicDescription> tag under <coverage>. Park Unit Connection(s) are the (typically) four-letter codes describing the park unit(s) where data were collected (e.g. ROMO, not ROMN). If there are already =items listed under geographicCoverage, Park Unit Connections will be inserted as the first item in the list of geographicCoverages and will be prefaced by the string "NPS Unit Connections:".
}
\examples{
ParkUnits<-("ROMO, GRSD, TRYME")
set.parkUnits(emlObject, ParkUnits)
}
