% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set_creator_order}
\alias{set_creator_order}
\title{Rearrange the order of creators (authors)}
\usage{
set_creator_order(eml_object, new_order = NA, force = FALSE, NPS = TRUE)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EML::read_eml(\if{html}{\out{<filename>}}, from="xml").}

\item{new_order}{List. Defaults to NA for interactively re-ordering the creators.}

\item{force}{logical. Defaults to false. If set to FALSE, a more interactive version of the function requesting user input and feedback. Setting force = TRUE facilitates scripting.}

\item{NPS}{Logical. Defaults to TRUE. \strong{Most users should leave this as the default}. Only under specific circumstances should it be set to FALSE: if you are \strong{not} publishing with NPS, if you need to set the publisher location to some place other than the Fort Collins Office (e.g. you are NOT working on a data package) or your product is "for" the NPS by not "by" the NPS and you need to specify a different agency, set NPS = FALSE. When NPS=TRUE, the function will over-write existing publisher info and inject NPS as the publisher along the the Central Office in Fort Collins as the location. Additionally, it sets the "for or by NPS" field to TRUE and specifies the originating agency as NPS.}
}
\value{
eml_object
}
\description{
The \code{set_creator_order()} function allows users to rearrange to order of creators (authors on DataStore).
}
\examples{
\dontrun{
meta2 <- set_creator_order(eml_object)}
}
