% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set_int_rights}
\alias{set_int_rights}
\title{Set Intellectual Rights}
\usage{
set_int_rights(
  eml_object,
  license = c("CC0", "public", "restricted"),
  force = FALSE,
  NPS = TRUE
)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EML::read_eml(\if{html}{\out{<filename>}}, from="xml").}

\item{license}{String. Indicates the type of license to be used. The three potential options are "CC0" (CC zero), "public" and "restricted". CC0 and public can only be used if CUI is set to either PUBFUL or PUBVER. Restricted can only be used if CUI is set to any code that is NOT PUBFUL or PUBVER (see \code{set_cui()} for a list of codes). To view the exact text that will be inserted for each license, please see https://nationalparkservice.github.io/NPS_EML_Script/stepbystep.html#intellectual-rights}

\item{force}{logical. Defaults to false. If set to FALSE, a more interactive version of the function requesting user input and feedback. Setting force = TRUE facilitates scripting.}

\item{NPS}{Logical. Defaults to TRUE. \strong{Most users should leave this as the default}. Only under specific circumstances should it be set to FALSE: if you are \strong{not} publishing with NPS, if you need to set the publisher location to some place other than the Fort Collins Office (e.g. you are NOT working on a data package) or your product is "for" the NPS by not "by" the NPS and you need to specify a different agency, set NPS = FALSE. When NPS=TRUE, the function will over-write existing publisher info and inject NPS as the publisher along the the Central Office in Fort Collins as the location. Additionally, it sets the "for or by NPS" field to TRUE and specifies the originating agency as NPS.}
}
\value{
emlObject
}
\description{
set_int_rights allows the intellectualRights field in EML to be surgically replaced.
}
\details{
set_int_rights requires that CUI information be listed in additionalMetadata prior to being called. The verbose \code{force = FALSE} option will warn the user if there is no CUI specified. set_int_rights checks to make sure the CUI code specified (see \code{set_cui()}) is appropriate for the license type chosen.
}
\examples{
\dontrun{
set_int_rights(eml_object, "CC0", force=TRUE, NPS=FALSE)
set_int_rights(eml_object, "restricted")}

}
