% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editEMLfunctions.R
\name{set_title}
\alias{set_title}
\title{Edit data package title}
\usage{
set_title(eml_object, data_package_title, NPS = TRUE)
}
\arguments{
\item{eml_object}{is an R object imported (typically from an EML-formatted .xml file) using EML::read_eml(<filename>, from="xml").}

\item{data_package_title}{is a character string that will become the new title for the data package. It can be specified directly in the function call or it can be a previously defined object that holds a character string.}

\item{NPS}{is a logical that defaults to TRUE. Checks EML for NPS publisher info and injects it if publisher is empty. If publisher already exists, it will be replaced by NPS (with publisher location set to the Fort Collins Office). If you are not publishing with NPS, set to FALSE.}
}
\value{
an EML-formatted R object
}
\description{
Edit data package title
}
\details{
The set_title function checks to see if there is an existing title and then asks the user if they would like to change the title. Some work is still needed on this function as get_eml() automatically returns all instances of a given tag. Specifying which title will be important for this function to work well.
}
\examples{
 \dontrun{
data_package_title<-"New Title. Must match DataStore Reference title."
eml_object<-set_title(eml_object, data_package_title)
}
}
